requirements.txt
requests
gitpython
autopep8
pylint